requirements.txt
# Core HTTP client (used by bot/client.py)
requests>=2.31.0

# Environment variable loading
python-dotenv>=1.0.0

# FastAPI backend (server.py)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
